upload project after finishing it
This is a project that generates a fake script from a provided script. 
There are no file trained_rnn.pt in this upload because the file is too large. In fact, that file is created during training process. It will be modified due to new data

Footnote: The project content and data were provided by Udacity machine learning course. I just finished it.